Decompile func_801027C4
