+++ config/splat.us.dra.yaml
@@ -50,6 +50,8 @@ segments:
       - [0x3C160, rodata]
       - [0x3CBC4, .rodata, 5D7C0]
       - [0x3CBDC, rodata]
+      - [0x3CC0C, .rodata, 627C4]
+      - [0x3CC2C, rodata]
       - [0x3CC80, rodata]
       - [0x3CD00, rodata]
       - [0x3CD10, rodata]
@@ -75,6 +77,7 @@ segments:
       - [0x46FD4, c, 46FD4]
       - [0x47BB8, c, 47BB8]
       - [0x5D7C0, c, 5D7C0]
+      - [0x627C4, c, 627C4]
       - [0x71830, c, 71830]
       - [0x73AAC, c, 73AAC]
       - [0x75DA0, c, 75DA0]
